refactor: add types for SNIP responses
